requirements.txt
Django==1.10.6
Markdown==2.6.6
Pillow==3.1.0
PyJWT==1.4.1
Pygments==2.1.3
argparse==1.2.1
backports.shutil-get-terminal-size==1.0.0
coreapi==1.32.3
decorator==4.0.10
dj-database-url==0.4.1
django-countries==4.0
django-filter==0.13.0
django-rest-swagger==2.0.5
djangorestframework==3.4.0
djangorestframework-jwt==1.8.0
drfdocs==0.0.11
gunicorn==19.6.0
httplib2==0.9.2
ipython==5.0.0
ipython-genutils==0.1.0
itypes==1.1.0
oauth2==1.9.0.post1
openapi-codec==1.0.0
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.3
psycopg2==2.6.1
ptyprocess==0.5.1
requests==2.11.1
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
traitlets==4.2.2
uritemplate==3.0.0
wcwidth==0.1.7
whitenoise==3.2
wsgiref==0.1.2
